Change instr_decode to find_instruction.
Instead of trying to decode the instruction from memory, it instead finds an instructino in an instruction bank by op_code. This removes 0xCB prefix handling from the function. The caller has to handle 0xCB prefixes, but this will actually be desired when we start implementing instruction dispatch by clock-tick, since fetching 0xCB will be a separate tick that needs to be handled outside of find_instruction.

This change also removes the .cb_prefix field from Instruction.

It's not needed. We can check if the instruction has an 0xCB prefix by checking whether it's in the cb_instructions bank.