Rename snprintf_instruction to format_instruction.
This disassociates it from snprintf, since their return values differ. It also adds INSTRUCTION_STR_MAX to get rid of guess work in allocating the buffer. Finally, the implementation is changed to stop relying on the return value of snprintf, which is somewhat more difficult to use than I originally thought (needs error and truncation checking).