Fail if get_reg8 or set_reg8 is called on REG_HL_MEM.
It is always an error to call get_reg8 or set_reg8 on REG_HL_MEM. However, many instructions encode REG_HL_MEM in the same way as the other 8-bit registers. This is accident prone. Care must be taken for each of these instructions to make sure to handle REG_HL_MEM separately. We should crash loudly if either of these functions is accidentally called on REG_HL_MEM.