Fix interrupt handling and the HALT bug.
I think in general, this actually simplifies the implementation. The CPU now tracks a state, which was formerly ExecResult: DONE, EXECUTING, INTERRUPTING, and HALTED.

When waking from HALTED a NOP is executed, which fetches the next instruction, and execution returns to normal.

When an interrupt is pending in the DONE state, the state switches to INTERRUPTING and we do the 5 cycles to enter the interrupt handler. Then execution returns to normal. When calling the interrupt, the PC is assumed to have been pre-incremented by the previous instruction's final step that set state to DONE, so the pushed PC for the return is PC-1.

There are tests for all four interesting cases of HALT:
1) IME=false, IE&IF=0 — halt and do a NOP on wake up to fetch IR=[PC++].
2) IME=false, IE&IF≠0 — "the HALT bug", second instruction read twice.
3) IME=true, IE&IF=0 — halt and do a NOP on wake up before checking interrutps.
4) IME=true, IE&IF≠0 — service interrupts; don't do HALT until IE&IF=0 (case #3).