Rename mem to data and addr to offset in instruction formatting.
The goal is to make it easier to grep for accesses to Gameboy's mem array so we can more easily ensure that accesses to it are done through functions like store and fetch that are aware of what is legal/illegal to access. And in the debugger, we want a hook to watch when memory is written.

Unit tests can send and read and instruction formatting code can read mem fieldss. They are not part of "the emulation". But code that is part of running the emulation should not access the array directly.