Start implementing a 9p library.
The fuse implementation for 9p isn't working for me. It's somewhat difficult to use the plan9port lib9p in a program that isn't all-in on plan9port. So we need to implement our own 9p library to communicate with Acme. Plus, this is a hobby project, and it's just more fun stuff to do.