Fix bug when closing due to a bad received message.
The bug was that the receive thread would mark the Client9p as closed, wait for everyone on the queue to finish, and would free the memory for the client. But the issue is that we shouldn't free the memory until the user explicitly calls close9p() to signal that they acknowledge that they are done with it. Since the recv_thread was freeing the memory, not close9p(), a subsequent call to close9p() after the receive thread closed due to a bad receive would then access free()d memory.

Instead, the recv_thread doesn't free up the Client9p, but just signals when it is done. close9p() waits for the queue and the recv_thread finish, then frees memory. Since the user called close9p() they acknowledge that they will no longer use the Client9p, and if they do, then it's their issue.