Convert lib9's I/O to use syscalls instead of FILE and friends.
The issue is that FILE's semantics include buffering and possibly seeking when a file is opened for r+ (read/write). Seeking is illegal on sockets, so concurrent read/write to the FILE gets into a situation that causes illegal seeks. Instead, we need to use read(2) and write(2) directly. This has been encapsulated into io.{c,h}.

An additional nicety here is that the wrappers for read and write set errstr9() for more consistent error messages.