Special handling for ROM and Prohibited memory fetch/store.
We break on reading prohibited memory. This is almost certainly due to a bug in the emulator.

Storing to prohibited memory is allowed, however. We just ignore it. It seems Super Mario Land writes to this when clearing OAM, probably because a bug in the game.

We currently break when writing to ROM. These writes are attempting to access the Memory Bank Controller, which is currently not implemented.