Move drawing to Acme to a separate thread.
This lets us separate the redraw framerate from the rate of the emulator. The emulator will do a vblank at 60Hz, but the redrawing is 30Hz.

Also I removed the two writes to the lcd win per re-draw: 1) setting addr to #0, and 2) font, clean, show. There's no need to do these for every frame. This seems to speed things up reasonably.